Improve code organization in scenario processing
