+++ src/core/index.ts
@@ -7,7 +7,6 @@ export * from './datetime'
 export * from './stream'
 export * from './task-queue'
 export * from './census-columns'
-export * from './census-intersection'
 export * from './census-buffer'
 export * from './choropleth'
 export * from './strings'
