+++ src/tl/index.ts
@@ -3,6 +3,7 @@ export * from './route'
 export * from './agency'
 export * from './departure'
 export * from './census'
+export * from './census-intersection'
 export * from './feed-version'
 export * from './jobs'
 export * from './departure-cache'
